refactor(domain): move domain to /utils
