+++ src/lib/series/domains/domain.ts
@@ -1,4 +1,4 @@
-import { Domain } from '../../utils/domain';
+import { Domain } from '../../../utils/domain';
 import { ScaleType } from '../../../utils/scales/scales';
 
 export interface BaseDomain {
