+++ src/state/utils.ts
@@ -40,7 +40,7 @@ import {
 import { ColorConfig, Theme } from '../lib/themes/theme';
 import { identity, mergePartial } from '../utils/commons';
 import { Dimensions } from '../lib/utils/dimensions';
-import { Domain } from '../lib/utils/domain';
+import { Domain } from '../utils/domain';
 import { AxisId, GroupId, SpecId } from '../utils/ids';
 import { Scale } from '../utils/scales/scales';
 import { SeriesDomainsAndData } from './chart_state';
